Tried to implement the Interface in c++
This is a lot more difficult than it should be I may need to refactor the way this will work but I do have the class in there and will work on the other stuff next. I am having issues with creating functions using the interaction interface in c++ it's causing errors.